+++ anvil_consortium_manager/views.py
@@ -254,20 +254,10 @@ class SingleAccountMixin(object):
 
 class AccountDetail(
     auth.AnVILConsortiumManagerViewRequired,
-    SingleTableMixin,
     SingleAccountMixin,
     DetailView,
 ):
     """Render detail page for an :class:`anvil_consortium_manager.models.Account`."""
-
-    context_table_name = "group_table"
-
-    def get_table(self):
-        """Get a table of :class:`anvil_consortium_manager.models.ManagedGroup` s that this account is a member of."""
-        return tables.GroupAccountMembershipTable(
-            self.object.groupaccountmembership_set.all(),
-            exclude=["account", "is_service_account"],
-        )
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
@@ -281,6 +271,19 @@ class AccountDetail(
         )
         context["show_deactivate_button"] = not context["is_inactive"]
         context["show_reactivate_button"] = context["is_inactive"]
+
+        context["group_table"] = tables.GroupAccountMembershipTable(
+            self.object.groupaccountmembership_set.all(),
+            exclude=["account", "is_service_account"],
+        )
+
+        workspace_sharing = models.WorkspaceGroupSharing.objects.filter(
+            workspace__in=self.object.get_accessible_workspaces()
+        ).order_by("workspace", "group")
+
+        context["accessible_workspace_table"] = tables.WorkspaceGroupSharingTable(
+            workspace_sharing
+        )
         return context